Extract leaderboard response rendering
